Use trait to bundle back button functionality in one file
